separated reporting server and client component
